Basic setup/home screen app
The app has 2 possible screens: setup and home. Whenever the app is opened for the first time it goes to the setup screen and after pressing the button the settings are saved on the phone and the screen changes to "home". So anytime the background settings have the variable "isSetupCompleted" as true, anytime the app is opened it goes directly to "home" screen. It is also possible to reset the settings by pressing the button in "home", which will make it go back to "setup".